+++ sdkproject/Assets/Mapbox/Unity/Location/UnityLocationWrappers/MapboxLocationInfoMock.cs
@@ -17,11 +17,11 @@ namespace Mapbox.Unity.Location
 
 		private Location _location;
 
-		public float latitude { get { return (float)_location.LatitudeLongitude.x; } }
+		public double latitude { get { return (float)_location.LatitudeLongitude.x; } }
 
-		public float longitude { get { return (float)_location.LatitudeLongitude.y; } }
+		public double longitude { get { return (float)_location.LatitudeLongitude.y; } }
 
-		public float altitude { get { return 0f; } }
+		public double altitude { get { return 0f; } }
 
 		public float horizontalAccuracy { get { return _location.Accuracy; } }
